Deploy script extension

+++ develop_deployment.sh
@@ -19,8 +19,10 @@ if [[ "$STARTUP_RESULT" -ne 0 ]]; then
   exit 1
 fi
 
-trap "exit" INT TERM ERR
-trap "kill 0" EXIT
+if [ "$1" != "skipTrap" ]; then
+  trap "exit" INT TERM ERR
+  trap "kill 0" EXIT
+fi
 
 SERVER_PORT=$RANDOM
 DATABASE_PORT=$RANDOM
